Skip tests on hint for update/delete for MongoDB < 4.4
