+++ docs/changelog.rst
@@ -15,7 +15,7 @@ Development
 - Further to the deprecation warning, remove ability to use an unpacked list to `Queryset.aggregate(*pipeline)`, a plain list must be provided instead `Queryset.aggregate(pipeline)`, as it's closer to pymongo interface
 - Further to the deprecation warning, remove `full_response` from `QuerySet.modify` as it wasn't supported with Pymongo 3+
 - Fixed stacklevel of many warnings (to point places emitting the warning more accurately)
-- Add support for collation/hint/comment to delete/update/aggregate #2842
+- Add support for collation/hint/comment to delete/update and aggregate #2842
 
 Changes in 0.29.0
 =================
